cloud: fix current file counter
Since b9e501dcd4e0 the number of started files is printed, but the wrong
counter is used. It worked on the single node case, but, when opening
files in multiple nodes, the counter would show:

    ... [1/N]
    ... [1/N]
    ... [1/N]
    ... [1/N]

instead of the expected:

    ... [1/N]
    ... [2/N]
    ... [3/N]
    ... [4/N]

Fix by using a counter for started files instead of using the processed
files counter.