cloud: fix filtering front-end DPE
The front-end DPE was always used because the name was being compared
with the IP of the reported DPE (the condition was false and thus the
DPE was not filtered).

Fix by comparing by full DPE name. Now the front-end DPE will not be
used unless is set by the command line options.