build.gradle: set CLARA as compile only dependency
Use native support for compile only dependencies instead of a workaround
with an extra configuration.

Need to extend the testCompile configuration with compileOnly though,
otherwise the tests will not find CLARA classes.

This also fixes deploying third-party dependencies (if the dependency
is also a dependency of CLARA, it was not copied to the lib/ directory).