base: fix default local node for multi-lang DPEs
For the threads using the DataManager service to stage files, create a
local reconstruction node object ignoring the other languages and just
use the Java DPE (i.e. the front-end), because the only required service
is the DataManager.

For the orchestrator using the local reconstruction node, request the
registered local DPES to the registrar, because they may be using
non-default ports.